requirements.txt
openai==1.12.0
cohere==4.13.1
anthropic==0.3.4
regex==2022.10.31
appdirs==1.4.4
jinja2==2.11.3
markupsafe==2.0.1
huggingface_hub==0.12
pytest==7.2.2
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
unifyai==0.9.2

